+++ spec/ruby_smart/simple_logger/extensions/helper_spec.rb
@@ -152,6 +152,12 @@ RSpec.describe "Helper extension" do
     end
   end
 
+  describe '#_tagged' do
+    it 'tags provided string' do
+      expect(@logger.send(:_tagged, 'Test', :hint)).to eq "[\e[46mHINT\e[0m] Test"
+    end
+  end
+
   describe '#_lgth' do
     it 'forces string to exact length' do
       expect(@logger.send(:_lgth, 'test', 10)).to eq "test      "
@@ -213,6 +219,28 @@ RSpec.describe "Helper extension" do
       expect{
         @logger.send(:_logdev, {}, :logdev)
       }.to raise_exception RuntimeError, "Unable to build SimpleLogger! The provided device 'logdev' must respond to 'write'!"
+    end
+  end
+
+  describe '#_scene_subject_with_opts' do
+    it 'returns default' do
+      expect(@logger.send(:_scene_subject_with_opts, [], 'Default')).to eq ['Default', {}]
+    end
+
+    it 'returns subject' do
+      expect(@logger.send(:_scene_subject_with_opts, ['Custom'], 'Default')).to eq ['Custom', {}]
+    end
+
+    it 'returns opts' do
+      expect(@logger.send(:_scene_subject_with_opts, [{lvl: :debug}], 'Default')).to eq ['Default',{lvl: :debug}]
+    end
+
+    it 'returns subject with opts' do
+      expect(@logger.send(:_scene_subject_with_opts, ['Custom', {lvl: :debug}], 'Default')).to eq ['Custom',{lvl: :debug}]
+    end
+
+    it 'returns skipped with opts' do
+      expect(@logger.send(:_scene_subject_with_opts, [nil, {lvl: :debug}], 'Default')).to eq ['Default',{lvl: :debug}]
     end
   end
 end